:sparkles: Fixes issue #2, Bruker data now work.
The number of scans in the input `.mz5` file now uses the
`'SpectrumIndex'` entry rather than the `'ChromatogramIndex'`
entry in `.mz5` files. Some vendor formats do not store a
total ion chromatogram which is required to generate the
`'ChromatogramIndex'` entry. Bruker TOF data files suffer from
this particular issue.

Some other changes account for empty scans, while this is not
a common occurrence, some MSn scans do not contain data. This
was previously breaking the `fill_lookup` function. A warning
message is printed when this is detected, the symptom being
a lack of minimum or maximum _m/z_ for a given scan.