Update pepFootGui.py
The boolean operators for comparing to `significance` and `threshold` when plotting in the Analysis Tab have been amended to be more intuative. Although in practice the calculated values of f_modand p-value rarely match the thresholds set exactly, in principle now significant labelling is defined as f_mod >= `threshold` and significant (un)masking is defined as p-value <= `significance`.